+++ src/modules/files/files.module.ts
@@ -1,8 +1,8 @@
 import { Module } from '@nestjs/common';
 import { FilesController } from './files.controller';
-import { MinioClientModule } from './../../storage/minio/minio.module';
 import { FilesService } from './files.service';
 import { HttpModule } from '@nestjs/axios';
+import { MinioClientModule } from '../../providers/minio/minio.module';
 
 @Module({
   imports: [MinioClientModule, HttpModule],
