+++ src/providers/minio/minio.module.ts
