+++ src/providers/minio/minio.service.ts
